2021 23 part 2
Solved with A*

Had a bug in what is now line 113. Had end_x instead of start_x, which made the whole route_safe function useless. Search-space got way too big, and it took forever to run. Now it runs in 10 seconds for both parts including test-case. I don't even understand how it worked for part1 with that bug.