feat: add functionality to save code snapshots when submitting
This is useful for when one wants a good deltatime, and therefore modify the part 1 code directly instead of using the part2 method. Often times, part 2 is not conceptually different from part 1, so it is usual to copy paste the part 1 code to the part2 method before starting work. Having a snapshot when first getting correct on a part ensures that one can find back to a working solution for both parts.